Add a default "region" argument for the three "aws" provider declarations in providers.tf to pass `pre-commit` tests (specifically the `terraform_validate` hook).
Per the Terraform documentation:
region - (Optional) This is the AWS region. It must be provided, but it can also
be sourced from the AWS_DEFAULT_REGION environment variables, or via a shared
credentials file if profile is specified.

So although these providers will be overriden by the calling module as seen in
"examples/basic_usage/main.tf", they need to have a value set to pass a
"terraform validate" check.