Remove dummy region from providers
terraform validate wants this, but adding any key inside the provider
block other than "alias" appears to actually _define_ a provider
instead of allowing one to be passed in from the parent module.

In this case, a provider is created from the default profile and the
specified region, and this is used _instead_ of the aws.images-staging
and aws.images-production providers passed in by the parent module.