Modify the role names for the SSM read roles created
Add `role_name` arguments to the two cisagov/ssm-read-role-tf-module
modules that indicated Production or Staging roles. This prevents
issues with duplicate resource names in a single account scenario and
actually mirrors what is done in the cisagov/ci-iam-user-tf-module
module that is used to create the CI user. The base value for the role
names is the current default for the `role_name` variable as defined
in the cisagov/ssm-read-role-tf-module configuration.